feat(nats): add nats package for publishing messages to a NATS server
This commit adds a new file `nats.go` in the `internal/message/nats`
directory. The file contains the implementation of a NATS publisher that
allows publishing messages to a NATS server. The `natsPublisher` struct
is defined, which holds an `EncodedConn` from the `nats-io/nats.go`
package. The `NewPublisher` function is implemented, which takes the
host and port of the NATS server as arguments and returns a new instance
of the `natsPublisher`. It establishes a connection to the NATS server
using the provided host and port, and creates an `EncodedConn` using the
`protobuf` encoder. The `Publish` method is implemented, which takes a
subject and a `Content` object as arguments and publishes the `Content`
object to the NATS server using the `EncodedConn`. The `Close` method is
implemented, which closes the `EncodedConn`.