feat(server.go): add server package and RunServer function to start the gRPC server
The server.go file is added to the internal/app/server directory to
contain the server package. The RunServer function is implemented to
start the gRPC server. Support for configurable logging format and level
is added to the getLogger function. The allParams function is added to
retrieve the ArangoDB connection parameters. The getGrpcOpt function is
added to retrieve the gRPC options for message topics. The
repoAndNatsConn function is added to establish the connection to the
ArangoDB repository and messaging server.