feat(arangodb.go): implement GetContentBySlug function to retrieve content by slug
The GetContentBySlug function has been implemented to retrieve content
from the database based on the provided slug. The function queries the
database using the ContentFindBySlug query and maps the slug parameter
to the query variables. If the response is empty, the function sets the
NotFound field of the returned model.ContentDoc to true. If there is an
error in getting the content or reading the response into the struct, an
appropriate error message is returned.