feat(arangodb.go): implement AddContent method to add new content to the database
The AddContent method has been implemented to add new content to the
database. It takes a NewContentAttributes struct as input and creates a
new ContentDoc model in the database with the provided attributes. The
method returns the created ContentDoc model or an error if there was an
issue creating the content.