test(arangodb_test.go): add unit test for adding content to the repository
The test case `TestAddContent` has been added to the `arangodb_test.go`
file. This test verifies the functionality of adding content to the
repository. It creates a new content object using the `NewStoreContent`
function and adds it to the repository using the `AddContent` method.
The test asserts that the returned content object has the expected name,
namespace, and slug values.