test(arangodb_test.go): add new test case TestGetContent to retrieve content by key
The new test case TestGetContent is added to the arangodb_test.go file
to test the functionality of retrieving content by key. This test case
ensures that the repository is able to retrieve content using the key as
an identifier. The implementation of the GetContent method in the
repository is added to support this functionality.