test(service_test.go): add test case for updating content
This commit adds a new test case `TestUpdateContent` to the
`service_test.go` file. The test case verifies the functionality of
updating content by making use of the `UpdateContent` method. It creates
a new content, updates it with new attributes, and then asserts that the
updated content has the expected values for attributes such as
`UpdatedBy`, `Content`, `UpdatedAt`, `Name`, `Namespace`, `Slug`, and
`CreatedBy`.