chore(ci.yml, lint.yml, testcov.yml): upgrade Go version and linter version
The Go version has been upgraded from 1.20 to 1.21 in ci.yml, lint.yml,
and testcov.yml to keep up with the latest stable release. This will
ensure that the project benefits from the latest features and
improvements in the Go language. The linter version has also been
upgraded to v1.55.2 in lint.yml for the same reasons.