ci(ci.yml): update Docker workflow to use docker-publish and add image and dockerfile details
The GitHub Actions CI workflow now uses the `docker-publish.yaml` from
the `main` branch instead of `docker-build-push.yaml` from the `develop`
branch, ensuring the use of the latest stable workflow. Additionally,
specifying the `image` and `dockerfile` parameters explicitly defines
the Docker image name and the location of the Dockerfile, which enhances
the clarity and control over the Docker build process in the CI
pipeline.