feat(go.mod): update go-genproto to latest version and add new dependencies
The go-genproto package is updated to a newer version to incorporate the
latest changes and improvements from the upstream repository.
Additionally, new dependencies, `github.com/emirpasic/gods` and
`github.com/jinzhu/now`, are added to enhance data structure handling
and provide better date-time utilities, respectively, which are
essential for the ongoing development and feature enhancements.